+++ BambooTracker/gui/instrument_selection_dialog.hpp
@@ -3,6 +3,7 @@
 #include <QDialog>
 #include <QVector>
 #include <memory>
+#include "configuration.hpp"
 
 class AbstractBank;
 
@@ -15,13 +16,28 @@ class InstrumentSelectionDialog : public QDialog
 	Q_OBJECT
 
 public:
-	InstrumentSelectionDialog(const AbstractBank &bank, const QString &text, QWidget *parent = nullptr);
-	~InstrumentSelectionDialog();
+	InstrumentSelectionDialog(const AbstractBank &bank, const QString &text,
+							  std::weak_ptr<Configuration> config, QWidget *parent = nullptr);
+	~InstrumentSelectionDialog() override;
 
 	QVector<size_t> currentInstrumentSelection() const;
 
+public slots:
+	void onJamKeyOnByMidi(int key);
+	void onJamKeyOffByMidi(int key);
+
+signals:
+	void jamKeyOnEvent(size_t id, JamKey key);
+	void jamKeyOnMidiEvent(size_t id, int key);
+	void jamKeyOffEvent(JamKey key);
+	void jamKeyOffMidiEvent(int key);
+
+protected:
+	bool eventFilter(QObject* watched, QEvent* event) override;
+
 private:
 	const AbstractBank &bank_;
+	std::weak_ptr<Configuration> config_;
 	std::unique_ptr<Ui::InstrumentSelectionDialog> ui_;
 
 	void setupContents();